# Conflicts:
#	src/main/java/com/textmessenger/repository/UserRepository.java

+++ src/main/java/com/textmessenger/repository/UserRepository.java
@@ -1,9 +1,10 @@
 package com.textmessenger.repository;
 
 import com.textmessenger.model.User;
+
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface UserRepository extends JpaRepository<User, Integer> {
+public interface UserRepository extends JpaRepository<User, Long> {
 }
